tests: consolidate openai_processing tests into tests/test_openai_processing.py
