--- tests/test_openai_processing_coverage.py
@@ -1,275 +1,0 @@
-# Comprehensive tests for openai_processing.py functionality
-import logging
-from unittest.mock import MagicMock, patch
-
-import pytest
-
-from src.conversation_manager import ThreadSafeConversationManager
-from src.openai_processing import process_openai_message
-
-
-class TestProcessOpenAIMessage:
-    @pytest.mark.asyncio
-    async def test_process_openai_message_success(self):
-        """Test successful OpenAI message processing."""
-        # Mock user
-        user = MagicMock()
-        user.id = 12345
-
-        # Mock conversation manager
-        conversation_manager = MagicMock(spec=ThreadSafeConversationManager)
-
-        # Mock OpenAI client
-        mock_response = MagicMock()
-        mock_response.choices = [MagicMock()]
-        mock_response.choices[0].message.content = "Test response from OpenAI"
-        mock_response.choices[0].finish_reason = "stop"
-        mock_response.id = "test-response-id"
-        mock_response.usage = MagicMock()
-        mock_response.usage.prompt_tokens = 50
-        mock_response.usage.completion_tokens = 20
-        mock_response.usage.total_tokens = 70
-
-        openai_client = MagicMock()
-        openai_client.chat.completions.create.return_value = mock_response
-
-        # Mock asyncio.to_thread
-        with patch("asyncio.to_thread") as mock_to_thread:
-            mock_to_thread.return_value = mock_response
-
-            result = await process_openai_message(
-                message="Hello OpenAI",
-                user=user,
-                conversation_summary=[],
-                conversation_manager=conversation_manager,
-                logger=logging.getLogger("test"),
-                openai_client=openai_client,
-                gpt_model="gpt-4",
-                system_message="You are a helpful assistant",
-                output_tokens=1000,
-            )
-
-            assert result == "Test response from OpenAI"
-            conversation_manager.add_message.assert_any_call(12345, "user", "Hello OpenAI")
-            conversation_manager.add_message.assert_any_call(
-                12345,
-                "assistant",
-                "Test response from OpenAI",
-                metadata={"ai_service": "openai", "model": "gpt-4"},
-            )
-
-            # Verify GPT-5 parameters were included
-            mock_to_thread.call_args[0][0]
-            # The call_args[0] is the lambda function, we need to execute it to check params
-            # This is a bit tricky to test directly, but we can verify the function was called
-
-    @pytest.mark.asyncio
-    async def test_process_openai_message_no_unsupported_params(self):
-        """Unsupported params are not sent and do not affect behavior."""
-        user = MagicMock()
-        user.id = 12345
-
-        conversation_manager = MagicMock(spec=ThreadSafeConversationManager)
-
-        mock_response = MagicMock()
-        mock_response.choices = [MagicMock()]
-        mock_response.choices[0].message.content = "Response"
-        mock_response.id = "test-id"
-
-        openai_client = MagicMock()
-        logger = logging.getLogger("test")
-
-        # Create a custom logger to capture log messages
-        with patch("src.openai_processing.logging.Logger.warning"):
-            with patch("asyncio.to_thread") as mock_to_thread:
-                mock_to_thread.return_value = mock_response
-
-                result = await process_openai_message(
-                    message="Test",
-                    user=user,
-                    conversation_summary=[],
-                    conversation_manager=conversation_manager,
-                    logger=logger,
-                    openai_client=openai_client,
-                    gpt_model="gpt-5",
-                    system_message="Test",
-                    output_tokens=1000,
-                )
-
-                assert result == "Response"
-                # The warning should be logged inside the lambda function
-                # This is harder to test directly, so we'll verify the response works
-
-    @pytest.mark.asyncio
-    async def test_process_openai_message_with_parameters(self):
-        """Test OpenAI processing with various parameters."""
-        user = MagicMock()
-        user.id = 12345
-
-        conversation_manager = MagicMock(spec=ThreadSafeConversationManager)
-
-        mock_response = MagicMock()
-        mock_response.choices = [MagicMock()]
-        mock_response.choices[0].message.content = "Response"
-        mock_response.id = "test-id"
-
-        openai_client = MagicMock()
-        logger = logging.getLogger("test")
-
-        with patch("asyncio.to_thread") as mock_to_thread:
-            mock_to_thread.return_value = mock_response
-
-            # Call without unsupported parameters
-            result = await process_openai_message(
-                message="Test",
-                user=user,
-                conversation_summary=[],
-                conversation_manager=conversation_manager,
-                logger=logger,
-                openai_client=openai_client,
-                gpt_model="gpt-5",
-                system_message="Test",
-                output_tokens=1000,
-            )
-
-            assert result == "Response"
-
-    @pytest.mark.asyncio
-    async def test_process_openai_message_empty_response(self):
-        """Test handling of empty OpenAI response."""
-        user = MagicMock()
-        user.id = 12345
-
-        conversation_manager = MagicMock(spec=ThreadSafeConversationManager)
-
-        # Mock empty response
-        mock_response = MagicMock()
-        mock_response.choices = [MagicMock()]
-        mock_response.choices[0].message.content = ""
-
-        openai_client = MagicMock()
-        logger = logging.getLogger("test")
-
-        with patch("asyncio.to_thread") as mock_to_thread:
-            mock_to_thread.return_value = mock_response
-
-            result = await process_openai_message(
-                message="Test",
-                user=user,
-                conversation_summary=[],
-                conversation_manager=conversation_manager,
-                logger=logger,
-                openai_client=openai_client,
-                gpt_model="gpt-4",
-                system_message="Test",
-                output_tokens=1000,
-            )
-
-            assert result is None
-
-    @pytest.mark.asyncio
-    async def test_process_openai_message_no_choices(self):
-        """Test handling of response with no choices."""
-        # Clear caches to ensure test isolation
-        from src.caching import conversation_cache, response_cache
-
-        conversation_cache.clear()
-        response_cache.cache.clear()
-
-        user = MagicMock()
-        user.id = 12345
-
-        conversation_manager = MagicMock(spec=ThreadSafeConversationManager)
-
-        # Mock response with no choices
-        mock_response = MagicMock()
-        mock_response.choices = []
-
-        openai_client = MagicMock()
-        logger = MagicMock()
-
-        with patch("asyncio.to_thread") as mock_to_thread:
-            mock_to_thread.return_value = mock_response
-
-            result = await process_openai_message(
-                message="Test",
-                user=user,
-                conversation_summary=[],
-                conversation_manager=conversation_manager,
-                logger=logger,
-                openai_client=openai_client,
-                gpt_model="gpt-4",
-                system_message="Test",
-                output_tokens=1000,
-            )
-
-            assert result is None
-            logger.error.assert_called()
-
-    @pytest.mark.asyncio
-    async def test_process_openai_message_timeout(self):
-        """Test handling of timeout errors."""
-        # Clear caches to ensure test isolation
-        from src.caching import conversation_cache, response_cache
-
-        conversation_cache.clear()
-        response_cache.cache.clear()
-
-        user = MagicMock()
-        user.id = 12345
-
-        conversation_manager = MagicMock(spec=ThreadSafeConversationManager)
-
-        openai_client = MagicMock()
-        logger = MagicMock()
-
-        with patch("asyncio.to_thread", side_effect=TimeoutError("Timeout")):
-            result = await process_openai_message(
-                message="Test",
-                user=user,
-                conversation_summary=[],
-                conversation_manager=conversation_manager,
-                logger=logger,
-                openai_client=openai_client,
-                gpt_model="gpt-4",
-                system_message="Test",
-                output_tokens=1000,
-            )
-
-            assert result is None
-            logger.exception.assert_called()
-
-    @pytest.mark.asyncio
-    async def test_process_openai_message_generic_exception(self):
-        """Test handling of generic exceptions."""
-        # Clear caches to ensure test isolation
-        from src.caching import conversation_cache, response_cache
-
-        conversation_cache.clear()
-        response_cache.cache.clear()
-
-        user = MagicMock()
-        user.id = 12345
-
-        conversation_manager = MagicMock(spec=ThreadSafeConversationManager)
-
-        openai_client = MagicMock()
-        logger = MagicMock()
-
-        with patch("asyncio.to_thread", side_effect=Exception("Generic error")):
-            result = await process_openai_message(
-                message="Test",
-                user=user,
-                conversation_summary=[],
-                conversation_manager=conversation_manager,
-                logger=logger,
-                openai_client=openai_client,
-                gpt_model="gpt-4",
-                system_message="Test",
-                output_tokens=1000,
-            )
-
-            assert result is None
-            logger.exception.assert_called()
-            # Should not add messages to conversation on failure
-            conversation_manager.add_message.assert_not_called()
